fix: shopping cart through only one js instead of 3 js
