--- services/shopping-cart/src/redisClient.js
@@ -1,13 +1,0 @@
-const { createClient } = require("redis");
-
-const client = createClient({
-  socket: {
-    host: process.env.REDIS_HOST || "redis",
-    port: parseInt(process.env.REDIS_PORT) || 6379,
-  },
-});
-
-client.on("error", (err) => console.error("Redis error:", err));
-client.connect();
-
-module.exports = client;
